refactor(api): move public API packages under api namespace
Move the transport-facing API packages (blockfrost, mesh, utxorpc) from
the module root into a dedicated api/ namespace so API code is grouped
separately from core node, ledger, database, networking, mempool, and
consensus packages.

- blockfrost -> api/blockfrost
- mesh       -> api/mesh
- utxorpc    -> api/utxorpc

Package names, routes, configuration, and server startup behavior are
unchanged; only import paths move. node.go imports are updated, and the
utxorpc tests that load immutable testdata via a relative path are
adjusted for the extra directory level. ARCHITECTURE.md package-location
references are updated to match.
