+++ api/utxorpc/sync_dump_history.go
